[#51] - Directorio Manager y refactor TwitchManager.php
